test: Fix reporting of assert associated with check_wheel_content
To allow pytest to display useful information, the tests containing
the helper function is registered.

See https://docs.pytest.org/en/latest/writing_plugins.html#assertion-rewriting
