+++ CHANGES.rst
@@ -31,6 +31,11 @@ Documentation
 Tests
 -----
 
+* Fix reporting of ``AssertionError`` raised in ``check_wheel_content()`` function by relocating the
+  source code into a dedicated module ``tests.pytest_helpers`` and by adding a ``conftest.py``
+  configuration file registering it for pytest assertion rewriting.
+  See https://docs.pytest.org/en/latest/writing_plugins.html#assertion-rewriting and :issue:`403`.
+
 * Fix ``test_generator_selection`` when building with "Visual C++ for Python 2.7"
   installed for all users. This addresses failure associated with ``win_c_compilervs2008cxx_compilervs2008python2.7``
   when running test in `scikit-build-feedstock <https://github.com/conda-forge/scikit-build-feedstock>`_ where
