+++ tests/__init__.py
@@ -11,18 +11,13 @@ import requests
 import six
 import subprocess
 import sys
-import wheel
 
 from contextlib import contextmanager
 from mock import patch
-from pkg_resources import parse_version
 
-from skbuild import __version__ as skbuild_version
 from skbuild.compat import which  # noqa: F401
 from skbuild.utils import push_dir
 from skbuild.platform_specifics import get_platform
-
-from zipfile import ZipFile
 
 
 SAMPLES_DIR = os.path.join(
@@ -292,48 +287,3 @@ def is_site_reachable(url):
         return request.status_code == 200
     except requests.exceptions.ConnectionError:
         return False
-
-
-def check_wheel_content(wheel_archive, expected_distribution_name, expected_content, pure=False):
-    """This function raises an AssertionError if the given wheel_archive
-    does not have the expected content.
-
-    Note that this function already takes care of appending the
-    ``<expected_distribution_name>.dist-info`` files to the ``expected_content``
-    list.
-    """
-
-    expected_content = list(expected_content)
-    expected_content += [
-        '%s.dist-info/top_level.txt' % expected_distribution_name,
-        '%s.dist-info/WHEEL' % expected_distribution_name,
-        '%s.dist-info/RECORD' % expected_distribution_name,
-        '%s.dist-info/METADATA' % expected_distribution_name
-    ]
-
-    if parse_version(wheel.__version__) < parse_version('0.31.0'):
-        expected_content += [
-            '%s.dist-info/DESCRIPTION.rst' % expected_distribution_name,
-            '%s.dist-info/metadata.json' % expected_distribution_name
-        ]
-
-    if pure:
-        assert wheel_archive.endswith('-none-any.whl')
-    else:
-        assert not wheel_archive.endswith('-none-any.whl')
-
-    archive = ZipFile(wheel_archive)
-    member_list = archive.namelist()
-
-    assert sorted(expected_content) == sorted(member_list)
-
-    # PEP-0427: Generator is the name and optionally the version of the
-    # software that produced the archive.
-    # See https://www.python.org/dev/peps/pep-0427/#file-contents
-    current_generator = None
-    with archive.open("%s.dist-info/WHEEL" % expected_distribution_name) as wheel_file:
-        for line in wheel_file:
-            if line.startswith(b"Generator"):
-                current_generator = line.split(b":")[1].strip()
-                break
-    assert current_generator == six.b("skbuild %s" % skbuild_version)
