+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest.register_assert_rewrite('tests.pytest_helpers')
