+++ tests/test_hello_cpp.py
@@ -18,8 +18,9 @@ from skbuild.utils import push_dir
 
 from zipfile import ZipFile
 
-from . import check_wheel_content, project_setup_py_test
+from . import project_setup_py_test
 from . import (_copy_dir, _tmpdir, SAMPLES_DIR)
+from .pytest_helpers import check_wheel_content
 
 
 def test_hello_builds():
