+++ tests/test_hello_pure.py
@@ -15,7 +15,8 @@ from skbuild.utils import push_dir
 
 from zipfile import ZipFile
 
-from . import check_wheel_content, project_setup_py_test
+from . import project_setup_py_test
+from .pytest_helpers import check_wheel_content
 
 
 @project_setup_py_test("hello-pure", ["build"], disable_languages_test=True)
